pyproject.toml
[tool.poetry]
name = "pyenv-poetry-example"
version = "0.1.0"
description = ""
authors = ["Gabriel Martin del Campo <gmartindelcye@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyleniumio = "^1.15.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
